fix: update builder pattern
