--- src/test/java/io/github/jinseisieko/bindecistree/BDTreeNodeTest.java
@@ -1,11 +1,0 @@
-package io.github.jinseisieko.bindecistree;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import org.junit.jupiter.api.Test;
-
-
-class BDTreeNodeTest {
-    @Test
-    void test() {
-        assertTrue(1 == 1);
-    }
-}
